Performance Improvement
Redesigned the algorithm to remove duplication that was causing severe
performance slow down. Instead of looping through the holdings in each
ETF, I created a central file that contains all the holdings for every
fund and removes any duplicates that are shared between funds. I then
grab the metrics for each holding. Once the metrics are grabbed I apply
the metrics to each individual ETF and it's holdings.